+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added ActionResult.
 - Added type hints functionality in action method parameters.
 - Added ActionResultException handling in action methods.
+- Session cookie is marked as "secure" if request is secure (https).
 - Deprecated AbstractActionResult.
 
 ## [2.0.1] 2018-06-19